+++ packages/js/admin-e2e-tests/src/pages/WcSettings.ts
@@ -44,7 +44,7 @@ export class WcSettings extends BasePage {
 
 	async paymentMethodIsEnabled( method = '' ): Promise< boolean > {
 		await this.navigate( 'checkout' );
-		await waitForElementByText( 'h2', 'Payment methods' );
+		await waitForElementByText( 'th', 'Method' );
 		const className = await getAttribute(
 			`tr[data-gateway_id=${ method }] .woocommerce-input-toggle`,
 			'className'
@@ -58,7 +58,7 @@ export class WcSettings extends BasePage {
 
 	async cleanPaymentMethods(): Promise< void > {
 		await this.navigate( 'checkout' );
-		await waitForElementByText( 'h2', 'Payment methods' );
+		await waitForElementByText( 'th', 'Method' );
 		const paymentMethods = await page.$$( 'span.woocommerce-input-toggle' );
 		for ( const method of paymentMethods ) {
 			if (