+++ packages/js/data/src/payment-gateways/index.ts
@@ -3,6 +3,7 @@
  */
 import { registerStore } from '@wordpress/data';
 import { controls } from '@wordpress/data-controls';
+import { SelectFromMap, DispatchFromMap } from '@automattic/data-stores';
 
 /**
  * Internal dependencies
@@ -12,6 +13,7 @@ import * as resolvers from './resolvers';
 import * as selectors from './selectors';
 import reducer from './reducer';
 import { STORE_KEY } from './constants';
+import { WPDataActions } from '../types';
 
 export const PAYMENT_GATEWAYS_STORE_NAME = STORE_KEY;
 
@@ -22,3 +24,13 @@ registerStore( STORE_KEY, {
 	controls,
 	reducer,
 } );
+
+declare module '@wordpress/data' {
+	// TODO: convert action.js to TS
+	function dispatch(
+		key: typeof STORE_KEY
+	): DispatchFromMap< typeof actions >;
+	function select(
+		key: typeof STORE_KEY
+	): SelectFromMap< typeof selectors > & WPDataActions;
+}