+++ packages/js/data/src/plugins/types.ts
@@ -2,6 +2,7 @@
  * Internal dependencies
  */
 import { pluginNames } from './constants';
+import { WPError } from '../types';
 
 export type RecommendedTypes = 'payments';
 
@@ -34,12 +35,17 @@ export type Plugin = {
 	content: string;
 	plugins: string[];
 	title: string;
+	category_additional: string[];
+	category_other: string[];
 	image: string;
+	image_72x72?: string;
 	square_image?: string;
 	recommendation_priority?: number;
 	is_visible?: boolean;
 	is_local_partner?: boolean;
 	is_offline?: boolean;
+	actionText?: string;
+	recommended?: boolean;
 };
 
 type PaypalOnboardingState = 'unknown' | 'start' | 'progressive' | 'onboarded';
@@ -56,3 +62,21 @@ export type PaypalOnboardingStatus = {
 		onboarded: boolean;
 	};
 };
+
+type PluginsResponse< PluginData > = {
+	data: PluginData;
+	errors: WPError< PluginNames >;
+	success: boolean;
+	message: string;
+} & Response;
+
+export type InstallPluginsResponse = PluginsResponse< {
+	installed: string[];
+	results: Record< string, boolean >;
+	install_time?: Record< string, number >;
+} >;
+
+export type ActivatePluginsResponse = PluginsResponse< {
+	activated: string[];
+	active: string[];
+} >;