+++ plugins/woocommerce-admin/bin/starter-pack/src/index.js
@@ -1,4 +1,0 @@
-// Import SCSS entry file so that webpack picks up changes
-import './index.scss';
-
-console.log( 'hello world' );