+++ plugins/woocommerce-admin/client/payments-welcome/cards/diners.js
@@ -1,4 +1,4 @@
-export default () => (
+export const DinersClub = () => (
 	/* eslint-disable */
 	<svg
 		width="51"