+++ plugins/woocommerce-admin/client/payments/payment-recommendations.tsx
@@ -13,7 +13,6 @@ import {
 	PAYMENT_GATEWAYS_STORE_NAME,
 	PLUGINS_STORE_NAME,
 	Plugin,
-	PluginsStoreActions,
 } from '@woocommerce/data';
 import { recordEvent } from '@woocommerce/tracks';
 import ExternalIcon from 'gridicons/dist/external';
@@ -24,6 +23,7 @@ import ExternalIcon from 'gridicons/dist/external';
 import './payment-recommendations.scss';
 import { createNoticesFromResponse } from '../lib/notices';
 import { getPluginSlug } from '~/utils';
+import { isWcPaySupported } from './utils';
 
 const SEE_MORE_LINK =
 	'https://woocommerce.com/product-category/woocommerce-extensions/payment-gateways/?utm_source=payments_recommendations';
@@ -41,7 +41,7 @@ const PaymentRecommendations: React.FC = () => {
 	const {
 		installAndActivatePlugins,
 		dismissRecommendedPlugins,
-	}: PluginsStoreActions = useDispatch( PLUGINS_STORE_NAME );
+	} = useDispatch( PLUGINS_STORE_NAME );
 	const { createNotice } = useDispatch( 'core/notices' );
 
 	const {
@@ -61,13 +61,16 @@ const PaymentRecommendations: React.FC = () => {
 					),
 				installedPaymentGateways: select( PAYMENT_GATEWAYS_STORE_NAME )
 					.getPaymentGateways()
-					.reduce( ( gateways, gateway ) => {
-						if ( installingGatewayId === gateway.id ) {
+					.reduce(
+						( gateways: { [ id: string ]: boolean }, gateway ) => {
+							if ( installingGatewayId === gateway.id ) {
+								return gateways;
+							}
+							gateways[ gateway.id ] = true;
 							return gateways;
-						}
-						gateways[ gateway.id ] = true;
-						return gateways;
-					}, {} ),
+						},
+						{}
+					),
 				isResolving: select( ONBOARDING_STORE_NAME ).isResolving(
 					'getPaymentGatewaySuggestions'
 				),
@@ -79,21 +82,11 @@ const PaymentRecommendations: React.FC = () => {
 		[ isInstalled ]
 	);
 
-	const supportsWCPayments =
-		paymentGatewaySuggestions &&
-		paymentGatewaySuggestions.filter( ( paymentGatewaySuggestion ) => {
-			return (
-				paymentGatewaySuggestion.id.indexOf(
-					'woocommerce_payments'
-				) === 0
-			);
-		} ).length === 1;
-
 	const triggeredPageViewRef = useRef( false );
 	const shouldShowRecommendations =
 		paymentGatewaySuggestions &&
 		paymentGatewaySuggestions.length > 0 &&
-		! supportsWCPayments &&
+		! isWcPaySupported( paymentGatewaySuggestions ) &&
 		! isDismissed;
 
 	useEffect( () => {
@@ -104,7 +97,7 @@ const PaymentRecommendations: React.FC = () => {
 		) {
 			triggeredPageViewRef.current = true;
 			const eventProps = ( paymentGatewaySuggestions || [] ).reduce(
-				( props, plugin ) => {
+				( props: { [ key: string ]: boolean }, plugin: Plugin ) => {
 					if ( plugin.plugins && plugin.plugins.length > 0 ) {
 						return {
 							...props,