+++ plugins/woocommerce/includes/admin/importers/class-wc-product-csv-importer-controller.php
@@ -91,12 +91,6 @@ class WC_Product_CSV_Importer_Controller {
 	 * @return bool
 	 */
 	public static function is_file_valid_csv( $file, $check_path = true ) {
-		/**
-		 * Can be used to override the decision to check the import file path.
-		 *
-		 * @param bool   $check_import_file_path If the import file path should be checked.
-		 * @param string $file                   Path of the file to be checked.
-		 */
 		return wc_is_file_valid_csv( $file, $check_path );
 	}