+++ packages/js/onboarding/CHANGELOG.md
@@ -1,7 +1,7 @@
 # Unreleased
 
 -   Update TaskList types.
-
+-   Added Typescript type declarations. #32615
 # 3.0.1
 
 -   Add missing dependency.